requirements.txt
dash
pandas
plotly
pyside6
sqlalchemy
pathvalidate
dash-bootstrap-components
pyarrow
#ADS1x15-ADC
hydra-core
python-dotenv
